+++ src/Utilities.cpp
@@ -1,13 +1,13 @@
 #include "./Utilities.h"
 
 namespace audio {
-	const double& Utilities::Clamp(const double& dValue, const double& dMin, const double& dMax) {
+	const float& Utilities::Clamp(const float& dValue, const float& dMin, const float& dMax) {
 		if (dValue < dMin) return dMin;
 		else if (dValue > dMax) return dMax;
 		return dValue;
 	};
 
-	const double& Utilities::Lerp(double& dValue, const double& dTarget, const double& dSpeed, const double& dMin, const double& dMax)
+	const float& Utilities::Lerp(float& dValue, const float& dTarget, const float& dSpeed, const float& dMin, const float& dMax)
 	{
 		// If the current value is not similar to the target value, add or subtract the factor.
 		if (dValue < dTarget - dSpeed || dValue > dTarget + dSpeed)
