+++ src/Utilities.h
@@ -4,7 +4,7 @@ namespace audio {
 	class Utilities final
 	{
 	public:
-		static const double& Clamp(const double& dValue, const double& dMin, const double& dMax);
-		static const double& Lerp(double& dValue, const double& dTarget, const double& dSpeed, const double& dMin, const double& dMax);
+		static const float& Clamp(const float& dValue, const float& dMin, const float& dMax);
+		static const float& Lerp(float& dValue, const float& dTarget, const float& dSpeed, const float& dMin, const float& dMax);
 	};
 }
